Initial implementation of showCoords app
This is the initial implementation of a simple Qt for Python application
that uses Qt Widgets to display an always-on-top window displaying the
current screen coordinates of the mouse pointer.

It is slightly intelligent in that it will reduce its update rate if the
mouse cursor does not move for some time.